# openapi.yaml
components:
  schemas:
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
    Get-api-send-inviteResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Get-api-send-inviteResponse.json
          format: uri
          readOnly: true
          type: string
      type: object
    LoginInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/LoginInputBody.json
          format: uri
          readOnly: true
          type: string
        password:
          type: string
        username:
          type: string
      required:
        - username
        - password
      type: object
    LoginOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/LoginOutputBody.json
          format: uri
          readOnly: true
          type: string
        token:
          type: string
      required:
        - token
      type: object
    Participant:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Participant.json
          format: uri
          readOnly: true
          type: string
        accessed:
          type: boolean
        email:
          type: string
        external_id:
          type: string
        id:
          format: int32
          type: integer
        name:
          type: string
        sent:
          type: boolean
        wa_number:
          type: string
      required:
        - id
        - external_id
        - name
        - email
        - wa_number
        - accessed
        - sent
      type: object
    Post-api-admin-attendanceRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Post-api-admin-attendanceRequest.json
          format: uri
          readOnly: true
          type: string
        participant_id:
          type: string
      required:
        - participant_id
      type: object
    Post-api-admin-attendanceResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Post-api-admin-attendanceResponse.json
          format: uri
          readOnly: true
          type: string
      type: object
    Post-api-admin-participantsRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Post-api-admin-participantsRequest.json
          format: uri
          readOnly: true
          type: string
        email:
          type: string
        name:
          type: string
        wa_number:
          type: string
      required:
        - name
        - email
        - wa_number
      type: object
    Post-api-admin-participantsResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Post-api-admin-participantsResponse.json
          format: uri
          readOnly: true
          type: string
      type: object
    Post-api-bulk-inviteResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - http://localhost:8080//schemas/Post-api-bulk-inviteResponse.json
          format: uri
          readOnly: true
          type: string
      type: object
info:
  description: API for managing wedding invitations, QR codes, and guest attendance
  title: Invite QR API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/admin/attendance:
    post:
      operationId: post-api-admin-attendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post-api-admin-attendanceRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post-api-admin-attendanceResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Post API admin attendance
      tags:
        - admin
  /api/admin/login:
    post:
      operationId: post-api-admin-login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Post API admin login
  /api/admin/participants:
    get:
      operationId: list-api-admin-participants
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Participant"
                type:
                  - array
                  - "null"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List API admin participants
      tags:
        - admin
    post:
      operationId: post-api-admin-participants
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post-api-admin-participantsRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post-api-admin-participantsResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Post API admin participants
      tags:
        - admin
  /api/bulk-invite:
    post:
      operationId: post-api-bulk-invite
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post-api-bulk-inviteResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Post API bulk invite
  /api/invite/{token}:
    get:
      operationId: get-api-invite-by-token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get API invite by token
  /api/qr:
    get:
      operationId: list-api-qr
      parameters:
        - explode: false
          in: query
          name: participant_id
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                contentEncoding: base64
                type: string
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List API qr
  /api/send-invite:
    get:
      operationId: get-api-send-invite
      parameters:
        - explode: false
          in: query
          name: guest_id
          schema:
            format: int32
            type: integer
        - explode: false
          in: query
          name: email
          schema:
            type: string
        - explode: false
          in: query
          name: wa_number
          schema:
            type: string
        - explode: false
          in: query
          name: name
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Get-api-send-inviteResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get API send invite
  /api/user:
    get:
      operationId: get-api-user
      parameters:
        - explode: false
          in: query
          name: id
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get API user
servers:
  - url: http://localhost:8080/